pyproject.toml
[project]
name = "gerador-fmv"
version = "0.1.0"
description = ""
authors = [
    {name = "Rodger",email = "rodgerdias6@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "customtkinter (>=5.2.2,<6.0.0)",
    "pillow (>=11.1.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
